pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ashoka-chatbot"
version = "0.1.0"
dependencies = [
    "Flask==2.3.3",
    "Flask-Cors==4.0.0",
    "sentence-transformers==2.2.2",
    "faiss-cpu==1.7.4",
    "openai==1.3.7",
    "gunicorn==21.2.0",
    "numpy==1.24.3",
    "torch==2.0.1"
]

[tool.setuptools.packages.find]
where = ["."]